requirements.txt
# Stream Processing
# ------------------------------------------------------------------------------
faust-streaming
faust-streaming[uvloop,prometheus]==0.6.4
pytz==2020.1
aiokafka==0.7.0
uvicorn==0.11.8  # https://github.com/encode/uvicorn
gevent==20.6.2
